serialization: replaced some int by uint64 to avoid ambiguity in serialization protocols

+++ backend/r1cs/r1cs.go
@@ -31,8 +31,8 @@ type R1CS interface {
 	io.WriterTo
 	io.ReaderFrom
 	IsSolved(solution map[string]interface{}) error
-	GetNbConstraints() int
-	GetNbWires() int
+	GetNbConstraints() uint64
+	GetNbWires() uint64
 	GetNbCoefficients() int
 }
